+++ DTCD-StyleSystem/src/base-components/components.js
@@ -7,6 +7,16 @@ import BaseInput from './BaseInput/BaseInput';
 import BaseSelect from './BaseSelect/BaseSelect';
 import GaugeSegmentBuilder from './GaugeSegmentBuilder/GaugeSegmentBuilder';
 import DataSourceSelect from './DataSourceSelect/DataSourceSelect';
+import BaseLabel from './BaseLabel/BaseLabel';
+import BaseLink from './BaseLink/BaseLink';
+import BaseIconButton from './BaseIconButton/BaseIconButton';
+import BaseChip from './BaseChip/BaseChip';
+import BaseTabs from './BaseTabs/BaseTabs';
+import BaseTooltip from './BaseTooltip/BaseTooltip';
+import BaseRadio from './BaseRadio/BaseRadio';
+import BaseRadioGroup from './BaseRadioGroup/BaseRadioGroup';
+import BaseRange from './BaseRange/BaseRange';
+import BaseExpandPanel from './BaseExpandPanel/BaseExpandPanel';
 
 export default [
   { name: 'base-button', baseClass: BaseButton },
@@ -16,6 +26,16 @@ export default [
   { name: 'base-color-picker', baseClass: BaseColorPicker },
   { name: 'base-input', baseClass: BaseInput },
   { name: 'base-select', baseClass: BaseSelect },
+  { name: 'base-label', baseClass: BaseLabel },
+  { name: 'base-link', baseClass: BaseLink },
+  { name: 'base-icon-button', baseClass: BaseIconButton },
+  { name: 'base-chip', baseClass: BaseChip },
+  { name: 'base-tabs', baseClass: BaseTabs },
+  { name: 'base-tooltip', baseClass: BaseTooltip },
+  { name: 'base-radio', baseClass: BaseRadio },
+  { name: 'base-radio-group', baseClass: BaseRadioGroup },
+  { name: 'base-range', baseClass: BaseRange },
+  { name: 'base-expand-panel', baseClass: BaseExpandPanel },
   { name: 'datasource-select', baseClass: DataSourceSelect },
   { name: 'gauge-segment-builder', baseClass: GaugeSegmentBuilder },
 ];